scripts: inst_cfg_mdb: Add die() function for better error handling
In the previous commit, we've removed subshell for the script. the
subshell along with -e makes the script to exit when any error occurs.

Instead, add a die function to handle errors more effectively by
providing specific error messages and exiting the script. This
approach replaces the need to encapsulate the entire script in a
subshell for error isolation.

Signed-off-by: Yasushi SHOJI <yashi@spacecubics.com>