py: create_xtce: Add --header option
Added the `--header` option to the `create_xtce` command to generate a
header file. This option can take a file path as an argument if you
want to specify a different file name.

Currently, `--header` is enabled by default, so the
`scripts/inst_cfg_mdb.sh` script does not need to be changed.

Signed-off-by: Yasushi SHOJI <yashi@spacecubics.com>