tctm: Add EPS File System Download
Added request and reply packets for downloading telemetry history.

However, the length of the entry data in the file download packet
varies depending on the type of telemetry, and cannot be identified by
the header part up to the command ID.

Currently, we have a common reply packet format with an entry data
length of up to 125 bytes, but the packet size does not match for
anything other than General Telemetry, causing a Yamcs warning.

In the future, we plan to update it so that it can be identified by
telemetry type.

Signed-off-by: Takahiro Inoue <takahiro.inoue@spacecubics.com>