tctm: Rename +/- parameter names
The following is defined according to the XTCE 1.2 Specification.

Parameter, and MetaCommand and other major entity names within this
database may be any length but may only contain numeric, a-z letters,
underscores, hyphens, or backslashes.

Contrary to this, we had used the symbols '+' and '-' in the parameter
names, so we will replace them with 'P' and 'M'.

Signed-off-by: Takahiro Inoue <takahiro.inoue@spacecubics.com>