tctm: Add shell command for Zero
Adds the shell command and reply telemetries for Zero.
The commands can accept strings up to 200 bytes in length, and the
command responses will be split into 240 byte segments when downlinked.
However, to avoid infinite downlinking, a limit is set on the number
of responses (50 responses per command).

Additionally, to account for potential command hang-ups, the timeout
duration (default:60sec) can also be specified from the Ground.

Signed-off-by: Takuya Sasaki <takuya.sasaki@spacecubics.com>