tctm: Add CRC save option
Since calculating CRC32 may take time, it might not be completed
within an AOS period. Therefore, an option will be added to specify
whether to store the CRC32 calculation result in the FRAM on SC-Sat1.
Only the result of the last CRC32 calculation will be stored in the
FRAM, and a command to retrieve this result will be added in the
next commit.

Signed-off-by: Takuya Sasaki <takuya.sasaki@spacecubics.com>