etc: Add Authentication settings
Until now, accessing Yamcs's web interface did not require
authentication. However, authentication is necessary for actual
operations, so we are adding the required settings.
In the test environment, requiring authentication every time would be
costly, so while the settings files will be added, it will set to
`disabled`.In the actual operations, you will need to set
`enabled: true` in `security.yaml`.

For primary authentication methods, we have added the
`YAML AuthModule`, which requires a username and password, however
the `password` field in `users.yaml` is left blank. Please input a
password generated using the PBKDF2 hash, following the instructions
in the documentation below:

https://docs.yamcs.org/yamcs-server-manual/security/authmodules/yaml/

And we are adding an `IP Address AuthModule` as a secondary
authentication method. This is to avoid the overhead of requiring
user and password authentication for scripts accessing Yamcs via the
Python API. By adding this module, we allow access without user and
password authentication only from the server where Yamcs is
installed. However, the `address` and `user` fields are left empty,
so please inputs them in according to actual environment.

For details on each authentication method, please refer to the
following documentation:

https://docs.yamcs.org/yamcs-server-manual/security/authmodules/

Signed-off-by: Takuya Sasaki <takuya.sasaki@spacecubics.com>