py: Add parameter check script
Adds a script to validate Yamcs data based on parameters and
comparison values defined in a YAML file. Currently, YAML
definitions are provided only for MAIN/EPS/SRS-3, which are active
during initial operations. The comparison values are also set to
what is expected to be obtained during the first AOS. This script
displays errors in the standard output when detected and saves all
validation results in a CSV file.

Signed-off-by: Takuya Sasaki <takuya.sasaki@spacecubics.com>