tctm: Fix EPS deploy message definition
The EPS has a feature to periodically downlink an arbitrary string
called a "deploy message." However, there are two issues with the
definition of this message:

  1. It belongs to the same container as CSP common function
  2. EPS adds a non-ASCII value, 0x01, to the first byte of the message

Regarding issue 2, it seems that a 1-byte command ID was
intentionally added, similar to other EPS telemetry. However, as a
result, the first byte of the `MESSAGE` parameter is not displayed
correctly.

This commit fixes the deploy message difinition to belong to the EPS
container and adjusts it to recognize the first byte as the command
ID.

Signed-off-by: Takuya Sasaki <takuya.sasaki@spacecubics.com>