etc: Add address field for HTTP server
In the actual operation, access to the Yamcs HTTP server will only be
allowed via VPN. Therefore, we are adding an `address` field to
explicitly specify the HTTP server's address. However, the default
value will be set to `localhost` for use in the testing environment.
In the actual operation, please configure the `address` and `port`
fields with values different from those used in the testing
environment.

Signed-off-by: Takuya Sasaki <takuya.sasaki@spacecubics.com>