tctm: Change in Session Duration for EPS File Download
With the current session duration setting for file downloads,
requesting the General Telemetry data for a large number of entries
could result in the Data Packet Stream being interrupted if the
session duration was exceeded.

To address this, the session duration has been adjusted to ensure
uninterrupted reception of up to the maximum (880) entries of General
Telemetry data.

Signed-off-by: Takahiro Inoue <takahiro.inoue@spacecubics.com>